+++ src/main/kotlin/org/wisp/gatesawakened/jumping/Jump.kt
@@ -32,30 +32,8 @@ internal object Jump {
         }
 
         // Jump player fleet to new system
-//        jumpPlayerToGate(destinationGate)
         di.sector.doHyperspaceTransition(di.sector.playerFleet, sourceLocation, JumpPointAPI.JumpDestination(destinationGate, null))
         return JumpResult.Success
-    }
-
-
-    private fun jumpPlayerToGate(gate: Gate) {
-        val playerFleet = di.sector.playerFleet
-        val newSystem = gate.starSystem
-
-        // Usable in the future?
-        // Global.getSector().doHyperspaceTransition(playerFleet, jumpPoint, dest);
-
-        // Jump player fleet to new system
-        playerFleet.containingLocation.removeEntity(playerFleet)
-        newSystem.addEntity(playerFleet)
-        di.sector.currentLocation = newSystem
-
-        // Move player fleet to the new gate's location
-        playerFleet.setLocation(gate.location.x, gate.location.y)
-
-        // Ensure that the player fleet's only action post-jump is to hang out around the gate
-        playerFleet.clearAssignments()
-        playerFleet.setMoveDestination(playerFleet.location.x, playerFleet.location.y)
     }
 
     sealed class JumpResult {